% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question.R
\docType{data}
\name{questions}
\alias{questions}
\title{Retrieve a interview question}
\format{
An object of class \code{character} of length 20.
}
\usage{
questions
}
\description{
Returns a interview question at random
}
\keyword{datasets}
