% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star.R
\name{star}
\alias{star}
\title{Interview STAR method}
\usage{
star()
}
\description{
Answering interview questions using the STAR method
'S' is situation of event where your experience took place.
'T' is task or problem you had to deal with and solve.
'A' is actions that were taken to solve task or problem
'R' is result of your actions, the total outcome
}
